+++ src/services/azdfs/backend.rs
@@ -494,7 +494,7 @@ impl AzdfsBackend {
     pub fn azdfs_update_request(
         &self,
         path: &str,
-        size: Option<u64>,
+        size: Option<usize>,
         body: AsyncBody,
     ) -> Result<Request<AsyncBody>> {
         let p = build_abs_path(&self.root, path);
